Fixed bug, changed stock source
The ESP chip occassionally stops getting data for an unknown reason, preventing time from updating. Fixed the error trapping logic so that if this occurs 3 times in a row, the PyPortal resets.

Also, Finnhub stopped providing index data in their free service. Replaced by tracking the SPY ETF, multiplying by 10, and adding 7. SPY is an ETF that tracks the S&P 500 and this gives a pretty close surrogate.